requirements.txt
streamlit==1.45.0
pandas==2.2.3
requests==2.32.0
openpyxl==3.1.5
python-dotenv==1.1.0
